// React/app/GerAdv_TS/AgendaQuem/Interfaces/interface.AgendaQuem.ts
// 100% auto-generated by Source Genesys WebApi Front & Back Creator
                     // copyright © 2000-2025 Menphis - Sistemas Inteligentes
                     // This file is part of the Source Genesys project                     
                    'use client';

export interface IAgendaQuem {
// 202501251
    id: number;
 
	advogado: number,
	funcionario: number,
	preposto: number,
	idagenda: number,
}

export interface IAgendaQuemFormProps {
  id: number;
  onClose: () => void;
  onSuccess: (registro?: any) => void;
  onError: () => void;
}

export interface IAgendaQuemIncProps {
    id: number;
    onClose: () => void;
    onSuccess: (registro?: any) => void;
    onError: () => void;
}